+++ astrocut/tess_footprint_cutout.py
@@ -34,6 +34,10 @@ class TessFootprintCutout(FootprintCutout):
         sequence number as an int or a list of sequence numbers. If not specified,
         cutouts will be generated from all sequences that contain the cutout.
         For the TESS mission, this parameter corresponds to sectors.
+    legacy_filenames : bool
+        If True, generated cutout filenames use the pre-1.2.0 format (``<ny>x<nx>`` size separator
+        and 6-decimal RA/Dec precision) instead of the current format (``<ny>-x-<nx>`` size separator
+        and 7-decimal RA/Dec precision). Default is False.
     verbose : bool
         If True, log messages are printed to the console.
 
@@ -72,9 +76,12 @@ class TessFootprintCutout(FootprintCutout):
         fill_value: Union[int, float] = np.nan,
         limit_rounding_method: str = "round",
         sequence: Union[int, List[int], None] = None,
+        legacy_filenames: bool = False,
         verbose: bool = False,
     ):
-        super().__init__(coordinates, cutout_size, fill_value, limit_rounding_method, sequence, verbose)
+        super().__init__(
+            coordinates, cutout_size, fill_value, limit_rounding_method, sequence, legacy_filenames, verbose
+        )
         # Make the cutouts upon initialization
         self.cutout()
 
@@ -122,6 +129,7 @@ class TessFootprintCutout(FootprintCutout):
             self._fill_value,
             self._limit_rounding_method,
             threads=8,
+            legacy_filenames=self._legacy_filenames,
             verbose=self._verbose,
         )